[CIR] Move complex lowering out of LoweringPrepare
The new pass was declared and defined inside LoweringPrepare.cpp, which left it
entangled with the pass it had just been split out of.  It moves to
ComplexLowering.cpp alongside the other CIR transforms, taking the mul and div
expansion with it, and LoweringPrepare keeps only the complex conj and cast
lowering it still owns.

`buildRuntimeFunction` was a file-static helper reached through a member
forwarder, which kept a generally useful routine private to one file.  It moves
to CIRTransformUtils, where the shared transform helpers live, and the call
sites in LoweringPrepare now call it directly so the forwarder is gone.
