+++ clang/lib/CIR/Dialect/Transforms/CMakeLists.txt
@@ -5,6 +5,7 @@ add_clang_library(MLIRCIRTransforms
   CIRCanonicalize.cpp
   CIRSimplify.cpp
   CIRTransformUtils.cpp
+  ComplexLowering.cpp
   CXXABILowering.cpp
   EHABILowering.cpp
   TargetLowering.cpp
